[add] localUserCredentials repository
